add allocated rms norm nki and verifications
