+++ .gitignore
@@ -81,3 +81,4 @@ Desktop.ini
 
 # Artifacts
 artifacts/
+private
